requirements.txt
requests==2.23.0
Django==3.0.8
jsonpickle==1.4.1
boto3==1.14.19
django-crispy-forms==1.9.1
django-storages==1.9.1
whitenoise==5.1.0
opencage==1.2.1
gunicorn==20.0.4 
django_heroku==0.3.1
Markdown==3.1.1
django-markdownx==3.0.1 
